# doc/openAPI.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Baby Statistic API",
    "description": "REST API for tracking baby milk, sleep, pee, and poop events.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local dev" }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/sleep": {
      "get": {
        "tags": ["Sleep"],
        "summary": "List all sleep records",
        "parameters": [
          { "$ref": "#/components/parameters/From" },
          { "$ref": "#/components/parameters/To" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Sleep" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Sleep"],
        "summary": "Create a sleep record",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SleepPost" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Sleep" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
...
    "/api/sleep/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "get": {
        "tags": ["Sleep"],
        "summary": "Get a sleep record by ID",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Sleep" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["Sleep"],
        "summary": "Update a sleep record",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SleepPut" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Sleep" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Sleep"],
        "summary": "Delete a sleep record",
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      },
      "From": {
        "name": "from",
        "in": "query",
        "required": false,
        "description": "Filter by created_at >= value (ISO datetime)",
        "schema": { "type": "string", "format": "date-time", "example": "2026-04-13T00:00:00" }
      },
      "To": {
        "name": "to",
        "in": "query",
        "required": false,
        "description": "Filter by created_at <= value (ISO datetime)",
        "schema": { "type": "string", "format": "date-time", "example": "2026-04-13T23:59:59" }
      }
    },
    "responses": {
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        },
        "required": ["error"]
      },
...
      "Timestamps": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "createdAt"]
      },
...
      "Sleep": {
        "allOf": [
          { "$ref": "#/components/schemas/Timestamps" },
          {
            "type": "object",
            "properties": {
              "start": { "type": "string", "format": "date-time", "example": "2026-04-13T22:00:00" },
              "end": { "type": "string", "format": "date-time", "nullable": true, "example": "2026-04-14T06:00:00" }
            },
            "required": ["start", "end"]
          }
        ]
      },
      "SleepPost": {
        "type": "object",
        "properties": {
          "start": { "type": "string", "format": "date-time", "example": "2026-04-13T22:00:00" },
          "end": { "type": "string", "format": "date-time", "nullable": true, "example": null }
        },
        "required": ["start"]
      },
      "SleepPut": {
        "type": "object",
        "properties": {
          "start": { "type": "string", "format": "date-time" },
          "end": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}